Moved camera into own file and added better light.
