+++ src/main.rs
@@ -5,6 +5,7 @@
 
 mod asset_tracking;
 mod audio;
+mod camera;
 mod demo;
 mod game;
 #[cfg(feature = "dev")]
@@ -17,12 +18,8 @@ use avian3d::prelude::*;
 use bevy::{
     asset::AssetMetaCheck,
     dev_tools::fps_overlay::FpsOverlayPlugin,
-    input::mouse::{AccumulatedMouseScroll, MouseScrollUnit},
-    prelude::*,
-    render::camera::ScalingMode,
+    prelude::*
 };
-
-const ZOOM_SCROLL_FACTOR: f32 = 256.;
 
 fn main() -> AppExit {
     App::new().add_plugins(AppPlugin).run()
@@ -73,6 +70,7 @@ impl Plugin for AppPlugin {
         app.add_plugins((
             asset_tracking::plugin,
             audio::plugin,
+            camera::plugin,
             game::plugin,
             demo::level::plugin,
             #[cfg(feature = "dev")]
@@ -96,17 +94,6 @@ impl Plugin for AppPlugin {
         // Set up the `Pause` state.
         app.init_state::<Pause>();
         app.configure_sets(Update, PausableSystems.run_if(in_state(Pause(false))));
-
-        // Spawn the main camera.
-        // TODO: Move camera and lighting stuff into the game folder.
-        app.add_systems(Startup, spawn_camera);
-        app.add_systems(Update, zoom_camera);
-        // TODO: Replace by directional light
-        app.insert_resource(AmbientLight {
-            color: Color::WHITE,
-            brightness: 800.,
-            ..default()
-        });
     }
 }
 
@@ -131,45 +118,3 @@ struct Pause(pub bool);
 /// A system set for systems that shouldn't run while the game is paused.
 #[derive(SystemSet, Copy, Clone, Eq, PartialEq, Hash, Debug)]
 struct PausableSystems;
-
-fn spawn_camera(mut commands: Commands) {
-    commands.spawn((
-        Name::new("Camera"),
-        Camera3d::default(),
-        Projection::from(OrthographicProjection {
-            // 6 world units per pixel of window height.
-            scaling_mode: ScalingMode::FixedVertical {
-                viewport_height: 32.0,
-            },
-            ..OrthographicProjection::default_3d()
-        }),
-        Transform::from_xyz(0.0, 15.0, 4.0).looking_at(Vec3::ZERO, Vec3::Y),
-    ));
-}
-
-fn zoom_camera(
-    projection: Single<&mut Projection>,
-    time: Res<Time>,
-    acc_scroll: Res<AccumulatedMouseScroll>,
-) {
-    let scroll_y = match acc_scroll.unit {
-        MouseScrollUnit::Pixel => acc_scroll.delta.y / 100.0,
-        MouseScrollUnit::Line => acc_scroll.delta.y,
-    };
-
-    let delta = scroll_y * ZOOM_SCROLL_FACTOR * time.delta_secs();
-
-    if delta == 0.0 {
-        return;
-    }
-
-    if let Projection::Orthographic(ref mut ortho) = *projection.into_inner() {
-        if let ScalingMode::FixedVertical { viewport_height } = &mut ortho.scaling_mode {
-            let autoscale_factor = 1. - (1.0 / (1. + *viewport_height));
-            *viewport_height += delta * autoscale_factor;
-            *viewport_height = viewport_height.clamp(8., 128.);
-
-            // info!(viewport_height, delta, scroll_y, acc_scroll.delta.y);
-        }
-    }
-}
