+++ src/screens/gameplay.rs
@@ -3,7 +3,7 @@
 use avian3d::prelude::{Physics, PhysicsTime};
 use bevy::{input::common_conditions::input_just_pressed, prelude::*, ui::Val::*};
 
-use crate::{Pause, demo::level::spawn_level, menus::Menu, screens::Screen};
+use crate::{Pause, menus::Menu, screens::Screen};
 
 pub(super) fn plugin(app: &mut App) {
     //app.add_systems(OnEnter(Screen::Gameplay), spawn_level);
